Refactor stellar object generation into modular classes
